// packages/uikit/packages/Frameworks/Dialog/Dialog.types.ts
export interface FunnelNavigationConfig {
  back?: boolean | ReactNode;
  backLocation?:
    | "header"
    | "footer"
    | { pc?: "header" | "footer"; mobile?: "header" | "footer" };
  next?: boolean | ReactNode;
  nextLocation?:
    | "header"
    | "footer"
    | { pc?: "header" | "footer"; mobile?: "header" | "footer" };
  exit?: boolean | ReactNode;
  exitLocation?:
    | "header"
    | "footer"
    | { pc?: "header" | "footer"; mobile?: "header" | "footer" };
}